requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
openai>=1.0
python-decouple>=3.8
gunicorn>=21.0
psycopg2-binary>=2.9
dj-database-url>=2.1
twilio>=9.0
